backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
sqlalchemy==2.0.25
pymongo==4.6.1
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
pillow==10.1.0